% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_tidy}
\alias{get_tidy}
\title{Returns a tidy-data version of a named dataset.}
\usage{
get_tidy(dataset)
}
\arguments{
\item{dataset}{The name of the dataset for which a tidy version
is requested. Available datasets include "anscombe".}
}
\value{
A tibble containing a tidied up version of the requested
  dataset.
}
\description{
Returns a tidy-data version of a named dataset.
}

