% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_tidy_airquality}
\alias{get_tidy_airquality}
\title{Creates a tidy-data version of the airquality dataset}
\usage{
get_tidy_airquality()
}
\description{
Dates are converted from Month / Day to lubridate
}

