% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_tidy_anscombe}
\alias{get_tidy_anscombe}
\title{Creates a tidy-data version of the Anscombe quartet dataset}
\usage{
get_tidy_anscombe()
}
\description{
Creates a tidy-data version of the Anscombe quartet dataset
}

